/api routes for taking nodes in and out of rotation: first up is HAProxy
We'll refine this a bit, then Elastic search.
Note: there's a critical race here where if a user hasn't visited the
nodes to load them up to poll, we're checking nothing on the first hit.
Changing everything to be an IModule or similar and begin polling on app
start is how I need things to behave for plugins anyway...so that's next
up.