Complete fix for #221
I'm giving up on Elasticsearch giving a fuck about breaking versions.
They don't. Breaks are intentionally not being constrained to major
versions, see Elastic #21410 and #21417 for context.

To support this reasonably, Opserver not just doesn't support pre-v0.90
correctly. If this becomes an issue we'll add it back quickly via a
switch.