MEGA GRAPH!
This is a first pass at a single SVG for ALL CPU using SVG view
fragments. It's not enabled because there are three big questionmarks:
- Can/should we switch based on Safari or not? (What the hell Apple,
update your mess)
- Fetching each interface is not ideal. Bosun for example has batch
fetching and this fires up a ton of unecessary Tasks to fetch from
ultimately the same in-memory data source.
- It's not output cached (due to result size) - we'll have to tweak
this.

The second leads to thread pool starvation for any sizable monitoring
set. I think the best way to resolve this is to have a
IBulkHistoryProvider or some such there the DataProvider has a few extra
Get____() routes for CPU, Memory, and Network history for efficiently
fetching in 1 instead of `n` queries. If that's present (and not
Safari?), we'd use this more optimal rendering which is FAR cheaper on
the client. And far fewer requests.