Exceptions overhaul
This overhauls how exceptions are done in total. All operations, querying, and paging is now done in SQL (which is also a step towards making it store agnostic). Many UI fixes and enhancements are involved as well, the big items are:
- Multiple stores supported
- Tabs for multiple stores
- More intuitive redirects after actions
- Consistent handling of Store, Group, and Log in the controller
- Fixes delete link in the detail page, and adds a protect link
- Removed ExceptionStores complexity, since operations are now per-tab and store specific...this just removes a layer of crazy for the 99% case.
- Allows customizing the Exceptional table name (available in Exceptional v2).

Note: this intentionally doesn't *require* Exceptional v2 features (e.g. new columns like Category), we'll do that when Exceptional v2 is out of RC.