Security simplification: first pass
I need to overhaul how all this works and cleaning it up to essentials first helps. More statics to remove and security to simplify next. I need to look at the provided auth providers and see what we can do here. While the current implementation is a Roles enum, this code swap should make it easy for them to be strings if needed later, e.g. ones defined in the module as their key for admin/viewer.