Another step in module cleanup
I haven't figure out exactly how NavTab should behave here, it needs the module and the controller to work, but maybe I can cleanly map to a [DefaultRoute] on the StatusController<T> where T : StatusModule in the status module registration section. This could register all tabs.

A reference from the controller to the module (non-transient) is ill-advised because due to config a module may come and go and we'd have ourselves a memory leak. It's better to have the module be a ref tree root and everything driven off that...I think. Still playing.

This compiles and runs, but top tabs are missing.