LocalCache -> IMemoryCache
Step 1 of moving to the built-in caching bits. Next up is giving Cache a ref to its PollNode which can see an instance of the PollingService, then we can eliminate the changes to SQL and such here for caches that need a backer, and not pass it down through every module...it'll just be on the PollingService the module's get in their constructor.

Note: this eliminates Current from Opserver.Core completely, yay simplification!