SignalFx: Use a different mechanism to get the list of available hosts
We're now using the list of hosts that have recently reported `cpu.utilization` which is what the SignalFx "Infrastructure" dashboards use. This uses a SignalFlow program that queries for hosts that reported values in the last 15 mins and drags the relevant metrics down so we can populate the relevant bits of the `Node` class.