Implement OpenId Connect security provider
This changes a few fundamentals of how the `SecurityProvider` model works. It makes the provider itself responsible for creating a `ClaimsPrincipal` that is then persisted using the usual ASP.NET Core authentication bits (in this case by a cookie).

Additionally, it adds support for different kinds of "token" that each provider validates in order to allow a user to login. E.g. the `ActiveDirectoryProvider` expects a `UsernamePasswordToken`, whereas the `OIDCProvider` expects an `OIDCToken` containing the JWT retrieved from the IdP.

The login flow itself is modified to support the different types of authentication expected by the various providers. If the provider expects `UsernamePassword` authentication then a username/password form is rendered, if the provider only requires a username (`EveryonesAnAdmin` or `EveryonesReadOnly`) then it renders just a username field and if it's an OIDC flow a login button is rendered that initiates the OIDC authorization flow.

Note that the username-only flow is a new concept, but fits with how `EveryonesAnAdmin` and `EveryonesReadOnly` work.

Settings for each provider are registered in the container and then, depending on the implementation requested the `SecurityManager` ensures the right settings are passed at runtime. A new abstract class `SecurityProvider<TSettings, TToken>` ensures that derived implementations get strongly-typed methods that they can operate against for token validation and accessing their settings.

Under the hood the OIDC flow redirects to the IdP's authorization URL and uses that to retrieve an access token. It extracts the ID token (represented as a JWT) from this access token and uses it to extract the relevant claims - notably the name of the user and the groups that they belong to. Okta, OneLogin and Azure all support sending groups in the ID token but Gsuite does not, making it effectively unsupported for now.

Finally, this adds documentation for configuring the OIDC provider and a smattering of unit tests to make sure it works properly.