SQL Module: Add support for monitoring SQL Azure instances
This PR adds support for connecting to SQL Azure server instances.

It extends `ISQLVersioned` to add a new property `ISet<SQLServerEdition> SupportedEditions` that is implemented by each data poller used by the `SQLModule`. When fetching properties for an instance a `SQLServerEdition` value is fetched that identifies the edition that the engine is running. When paired with the version we can adjust SQL and dynamically configure the supported pollers for running against SQL Azure. `SQLModule.GetFetchSQL` and `SQLModule.Supports` have been modified to take a `SQLServerEngine` struct that represents the combination of the version and the edition. Note: this likely needs some work to allow the version _and_ edition combination to factor into the decision on what SQL to run - it currently only handles one or the other.

Majority of differences in this PR are because the `msdb` database and some DMVs present on an in-premise installation of SQL Server are not present in SQL Azure. As a result most changes are adding the new `ISQLVersioned` property and then implementing the relevant branches in each data poller.

However, there is also the added wrinkle that just connecting to the `master` database on a SQL Azure instance does not allow connection to other databases hosted on that "server". In SQL Azure each DB is actually its own server instance. To support this I've added a new class `SQLAzureServer` that derives from `SQLInstance` and polls for the databases within the server. Each database that it finds is dynamically added as a `SQLInstance` and can be monitored as any normal server instance would be.