Fix AuthController.RedirectToProvider not round-tripping returnUrl
The state parameter is expected to decode as if it was a query string, but the values within the parameter weren't being encoded as query string values, so the decoding was lossy and would not round-trip the return URL if that URL contained more than one query parameter (e.g. for an exception detail URL: `/exceptions/detail?id=...&log=...&store=...`).