Update SQLInstance.Databases.cs
The important changes here are for performance, but there are other benefits as well.

Looping/cursor to execute DBCC per database is expensive, as are two #temp tables used to hold/pass results. This is pronounced on servers with a large number of databases.

The sysadmin requirement of the DBCC call is prohibitive (and as shown in issue 306 means those without sysadmin get 0s).

Since 2016 SP2, we've been able to extract VLF information from a single function call instead of looping over DBCC checks. The function only requires VIEW SERVER STATE, which is more likely to be granted in a lot of environments.

This commit checks for those lower permissions and, if we are 2016 SP2+, performs a single call that eliminates the cursor, both temp tables, and the sysadmin requirement.

If the object check fails (< 2016 SP2), OR the permissions check fails, then they fall through to the existing cursor/#temp table solution. That still may be an issue if permissions are inadequate, but it doesn't deviate from the current behavior (VLFs full of 0s).

I also changed the cursor to a cursor variable, which can have better default performance and also doesn't need close/deallocate.